Adding hist and trace to the pickle dump
Can you derive hist and trace from the approx object?  If not, then how does `hackett.ipynb` extract the `trace` and `hist` information without dumping it into a pickle?

The `approx` ADVI object does contain a `hist` attribute, but it is a `Tensor` object that does not contain a `sample()` method.